Image_auto_scoring/requirements.txt
pandas==2.2.2
Pillow==10.3.0
opencv-python-headless==4.10.0.84
ollama
tqdm
scipy

